simpler aboutView rather than settingsView
